+++ src/minFieldValue.js
@@ -1,13 +1,9 @@
-import validator, { isIgnored } from './validator';
+import validator from './validator';
 import { gte } from 'lodash';
 
 export default function minFieldValueValidator(field, max = 0, props) {
     props = Object.assign({}, props);
     props.message = props.message || `${field} no less than ${max}`;
-
-    props.isIgnored = props.isIgnored || ((value) => {
-        return isIgnored(value) || isIgnored(value[field]);
-    });
 
     return validator(
         (value) => gte(value[field], max),