+++ src/required.js
@@ -1,4 +1,7 @@
 import validator from './validator';
+import { isArray, isObject, isDate, isEmpty, isEqual } from 'lodash';
+
+const emptyDate = new Date(0);
 
 export default function requiredValidator(props) {
     props = Object.assign({}, props);
@@ -7,8 +10,21 @@ export default function requiredValidator(props) {
     // The required validator doesn't ignore any values by default
     props.isIgnored = props.isIgnored || (() => false);
 
-    return validator(
-        (value) => !!value,
-        props
-    );
+    const isValid = (value) => {
+        if (isArray(value)) {
+            return !isEmpty(value);
+        }
+
+        if (isDate(value)) {
+            return !isEqual(value, emptyDate);
+        }
+
+        if (isArray(value) || isObject(value)) {
+            return !isEmpty(value);
+        }
+
+        return !!value;
+    };
+
+    return validator(isValid, props);
 }