+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Added field validators and updated min/max validator messages.
 Added an isIgnored property for validator to override the decision for ignoring values; ValidationResult instances then get an isIgnored property to surface whether or not the validator was ignored.
 Utilize isIgnored so that required can use validator.
 Prevent mutation of props passed to validators.
+Change behavior of required and how empty values are identified.
 
 * minvalue renamed to minValue
 * maxvalue renamed to maxValue
@@ -16,6 +17,8 @@ Prevent mutation of props passed to validators.
     * fieldValue(field, min, max) validates the field is within a range
 * minFieldValue for validating a field's value is at least the min
 * maxFieldValue for validating a field's value is at most the max
+* required now treats empty arrays and objects as invalid
+    * All other validators check the result of required to determine if the value should be ignored
 
 ## 0.0.4
 API and implementation changes for length and value validators.